Corrected DBDatabase hierarchy so that DBDatabase is the interface and DBDatabaseImplementation is the concrete implementation of the mythical standard Added backticks to MySQL column and table names Upgraded MySQL to the version 8 driver and switched to the new driver class new
