+++ src/main/java/nz/co/gregs/dbvolution/databases/JavaDB.java
@@ -34,7 +34,7 @@ import nz.co.gregs.dbvolution.exceptions.ExceptionDuringDatabaseFeatureSetup;
  *
  * @author Gregory Graham
  */
-public class JavaDB extends DBDatabase {
+public class JavaDB extends DBDatabaseImplementation {
 
 	public static final String DRIVER_NAME = "org.apache.derby.jdbc.ClientDriver";
 	public static final long serialVersionUID = 1l;
