+++ src/main/java/nz/co/gregs/dbvolution/databases/MariaClusterDB.java
@@ -31,7 +31,7 @@ import nz.co.gregs.dbvolution.exceptions.ExceptionDuringDatabaseFeatureSetup;
  *
  * @author Gregory Graham
  */
-public class MariaClusterDB extends DBDatabase {
+public class MariaClusterDB extends DBDatabaseImplementation {
 
 	public final static String MARIADBDRIVERNAME = "com.mariadb.jdbc.Driver";
 	public static final long serialVersionUID = 1l;
