+++ src/main/java/nz/co/gregs/dbvolution/databases/PostgresDBOverSSL.java
@@ -154,10 +154,6 @@ public class PostgresDBOverSSL extends PostgresDB {
 		return super.clone(); //To change body of generated methods, choose Tools | Templates.
 	}
 
-//	@Override
-//	protected Class<? extends DBDatabase> getBaseDBDatabaseClass() {
-//		return PostgresDBOverSSL.class ;
-//	}
 	@Override
 	public PostgresOverSSLSettingsBuilder getURLInterpreter() {
 		return new PostgresOverSSLSettingsBuilder();
